backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-cloud-speech==2.18.0
google-cloud-storage==2.8.0
pydantic==1.10.9
SQLAlchemy==2.0.20
gunicorn==20.1.0
ask-sdk-core==1.16.1
ask-sdk-model==1.19.0
flask-ask-sdk==1.0.0
Flask-Limiter==3.3.1
adb-shell==0.3.2
Pillow>=10.0.0
